Got the code working so that it can create an SFM file from a csv with headers.
Got the code working so that it can create an SFM file from a csv with headers.
However there's a hard coded If True at the moment to  force that code to run.
It would be good to have an option to switch between types of csv file. Also it would be good to automatically ignore empty columns at the end.